backend_django/requirements.txt
asgiref==3.3.1
boto3==1.17.7
botocore==1.20.12
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
jmespath==0.10.0
jwcrypto==0.8
oauth2client==4.1.3
olefile==0.46
Pillow==8.1.0
pip==20.3.3
protobuf==3.15.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
Pyrebase4==4.4.3
PySocks==1.7.1
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.1
s3transfer==0.3.4
setuptools==52.0.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
wheel==0.36.2
whitenoise==5.2.0
